move API docs downwards
Sphinx 5.3 (compared to 5.1.1) is now putting all the autodoc
names into the TOC.   So we have to start being more careful
to make sure API docs are well below narrative docs, because
this new style is a wall of text.   i dont yet see any options
to turn it off, but it does seem like a good improvement, just makes
doc organization a more difficult endeavor.

Change-Id: I49428076fef9b96ef1544621de9a9dfca1699dab