evaluate types in terms of the class in which they appear
Fixed issues within the declarative typing resolver (i.e. which resolves
``ForwardRef`` objects) where types that were declared for columns in one
particular source file would raise ``NameError`` when the ultimate mapped
class were in another source file.  The types are now resolved in terms
of the module for each class in which the types are used.

Fixes: #8742
Change-Id: I236f94484ea79d47392a6201e671eeb89c305fd8